initial change to use the LTFS LE API
This commit is not a final integration of the changes to use the LTFS LE
API. It contains minimal changes to make it working. The following
changes are still required:

- completely rework the LEControl code since this is based on EE code
- additional messages and catching exceptions are required or need to be
reworked when calling LTFS LE API functions.

To use the LTFS LE API code a corresponding LTFS LE release needs to be
installed containing the API. The LD_LIBRARY_PATH environment variable
needs to be set accordingly.